% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.r
\docType{data}
\name{aggregates}
\alias{aggregates}
\title{\code{dgo} example data: aggregated item responses}
\format{A {data.frame} with 4,284 observations of 7 variables.}
\usage{
aggregates
}
\description{
A table of aggregated item responses given by survey respondents and their
characteristics. Data are from the General Social Survey.
}
\examples{
aggregates
}
\seealso{
\url{http://www.gss.norc.org}
}
