% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_state_race_targets.r
\docType{data}
\name{annual_state_race_targets}
\alias{annual_state_race_targets}
\title{\code{dgo} example data: annual U.S. population targets by state and race}
\format{A {data.frame} with 13,158 observations of 4 variables.}
\usage{
annual_state_race_targets
}
\description{
A table giving U.S. population proportions by year for combinations of state
and three-category race variables. These are the same data as in
\code{\link{targets}}, but aggregated over all the other demographic
variables in that dataset. For use with the \code{target_data} argument of
function \code{shape} in package '\href{https://jdunham.io/dgo/}{dgo}'. Data
are from the U.S. Census. Proportions sum to to 1 within years.
}
\examples{
annual_state_race_targets
}
\seealso{
\url{http://www.census.gov/}
}
