% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.r
\docType{data}
\name{states}
\alias{states}
\title{\code{dgo} example data: state demographics}
\format{A \code{data.frame} with 4,386 observations of 12 variables.}
\usage{
states
}
\description{
A table giving demographic characteristics of American states 1930-2015.
}
\examples{
states
}
