% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.r
\docType{data}
\name{targets}
\alias{targets}
\title{\code{dgo} example data: U.S. population targets}
\format{A \code{data.frame} with ~803K observations of 10 variables.}
\usage{
targets
}
\description{
A table giving U.S. population proportions by year for combinations of
demographic variables. For use with the \code{target_data} argument of
function \code{shape} in package '\href{https://jdunham.io/dgo/}{dgo}'. Data
are from the U.S. Census.  Proportions sum to to 1 within years.
}
\seealso{
\url{http://www.census.gov/}
}
